refactor: reorganise dirs and colocate utils
