+++ src/WorkerProcess/htmlTemplates/insertIntoDashboardHeaderTemplate.ts
