+++ src/WorkerProcess/htmlTemplates/insertIntoImageTemplate.ts
