+++ src/WorkerProcess/scrapeUtils/base64EncodeFile.ts
