+++ src/WorkerProcess/scrapeUtils/clearDir.ts
