+++ src/WorkerProcess/scrapeUtils/getDashboardItemVisualization.ts
@@ -6,7 +6,7 @@ import type {
     EventReport,
     EventVisualization,
     Visualization,
-} from '../types'
+} from '../../types'
 
 export function getDashboardItemVisualization(
     dashboardItem: DashboardItem
