+++ src/WorkerProcess/scrapeUtils/getNestedPropertyValue.ts
