+++ src/index.ts
@@ -1,31 +1,3 @@
-import cluster from 'node:cluster'
-import http from 'node:http'
-import process from 'node:process'
-import { PrimaryProcess } from './MainProcess'
-import { WorkerProcess } from './WorkerProcess'
-import { readEnv } from './utils'
-
-const initializeCluster = async () => {
-    const env = readEnv()
-
-    if (cluster.isPrimary) {
-        const primaryProcess = new PrimaryProcess(env)
-
-        http.createServer(primaryProcess.requestListener).listen(
-            env.port,
-            parseInt(env.host),
-            () => {
-                console.log(
-                    `DHIS2 Push Analytics Service is running on http://${env.host}:${env.port}`
-                )
-            }
-        )
-    } else {
-        console.log(
-            `Starting dashboard-item conversion worker on PID ${process.pid}`
-        )
-        await WorkerProcess.create(env, false)
-    }
-}
+import { initializeCluster } from './Cluster'
 
 initializeCluster()
