--- src/utils/createTimer.ts
@@ -1,6 +1,0 @@
-export function createTimer() {
-    const startTimestamp = Date.now()
-    const getElapsedTime = () =>
-        ((Date.now() - startTimestamp) / 1000).toFixed(2)
-    return { getElapsedTime }
-}
