--- src/utils/index.ts
@@ -1,18 +1,0 @@
-// Scraper/puppeteer related (AppScraper / Converter / ItemParser)
-export { base64EncodeFile } from './base64EncodeFile'
-export { clearDir } from './clearDir'
-export { createPuppeteerBrowser } from './createPuppeteerBrowser'
-export { createTimer } from './createTimer'
-export { downloadPath } from './downloadPath'
-export { getDashboardItemVisualization } from './getDashboardItemVisualization'
-export { getNestedPropertyValue } from './getNestedPropertyValue'
-export { logDashboardItemConversion } from './logDashboardItemConversion'
-export { validateRequest } from './validateRequest'
-export { waitForFileToDownload } from './waitForFileToDownload'
-
-// Only in 3 files, consider colocation
-export { getThreadLength } from './getThreadLength' // PrimaryProcess
-export { parseQueryString } from './parseQueryString' // RequestHandler
-
-// valid utils
-export { readEnv } from './readEnv'
